🐛 Fix remaining preset references causing build errors
- Updated templateStore to use template-converter imports
- Fixed app/page.tsx import
- Renamed preset-registry to template-registry
- Updated all function names and imports
- Fixed visualization-generators to use executeTemplate
- Added backward compatibility for PRESET_ID comment marker
